requirements.txt
instaloader>=4.10.3
requests>=2.31.0
python-dotenv>=1.0.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0
fake-useragent>=1.4.0
schedule>=1.2.0
colorama>=0.4.6
tqdm>=4.66.0
redis>=5.0.0
psycopg2-binary>=2.9.7
psutil>=5.9.0
prometheus-client>=0.17.0
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
